Ading missing package
According to the Ansible docs this package is required - I haven't noticed any issues so far by not including it but I don't really want to have to diagnose something were this would be the cause either.

Since the package is tiny I see no problems with this as an extra dependency.